Refactoring evaluation and tooling support into own modules
