+++ lib/java-service-installer.js
@@ -6,28 +6,28 @@ const request = require('request-promise-native');
 const path = require('path');
 
 function getJavaCoords(serviceName) {
-    const coords = {
-        groupId: packagejson[serviceName].groupId,
-        artifactId: packagejson[serviceName].artifactId,
-        version: packagejson[serviceName].version,
-        classifier: packagejson[serviceName].classifier
-      };
+  const coords = {
+    groupId: packagejson[serviceName].groupId,
+    artifactId: packagejson[serviceName].artifactId,
+    version: packagejson[serviceName].version,
+    classifier: packagejson[serviceName].classifier
+  };
 
-    return coords;
+  return coords;
 };
 
 function getJarHome() {
-    return path.join(__dirname, '..', 'jars');
+  return path.join(__dirname, '..', 'jars');
 }
 
-function getJarPath(serviceName) {
+function getServicePath(serviceName) {
   const jarHome = getJarHome();
   const jarName = getLocalName(getJavaCoords(serviceName));
   return path.join(jarHome, jarName);
 }
 
 function getLocalName(coords) {
-    return (coords.artifactId + "-" + coords.version + (coords.classifier ? ("-" + coords.classifier) : '') + '.jar')
+  return (coords.artifactId + "-" + coords.version + (coords.classifier ? ("-" + coords.classifier) : '') + '.jar')
 }
 
 function getSearchUrl(groupId, artifactId, version, classifier) {
@@ -45,25 +45,25 @@ async function installServiceIfRequired(serviceName, updateInstallCallback) {
 }
 
 async function isServiceInstalled(serviceName) {
-  let jarPath = getJarPath(serviceName);
+  let jarPath = getServicePath(serviceName);
   return fileExists(jarPath);
 }
 
 // Installs a jar using maven coordinates
 async function installJar(serviceName, updateInstallCallback) {
   let coords = getJavaCoords(serviceName);
-  let jarPath = getJarPath(serviceName);
+  let jarPath = getServicePath(serviceName);
   let jarHome = getJarHome();
 
   var doesExist = await fileExists(jarHome);
   if (!doesExist) {
-    fs.mkdirSync(home, { recursive: true });
+    fs.mkdirSync(jarHome, { recursive: true });
   }
 
   await updateInstallCallback(`installing ${serviceName}`);
   const setupInfo = await setupDownload(getSearchUrl(coords.groupId, coords.artifactId, coords.version, coords.classifier), serviceName, updateInstallCallback);
   await DownloadFile(setupInfo.serverDownloadUrl, jarPath, setupInfo.provideInstallStatus, setupInfo.serverDownloadSize);
- 
+
   doesExist = await fileExists(jarPath);
   if (!doesExist) {
     throw Error(`Failed to install ${serviceName}`);
@@ -120,13 +120,5 @@ function handleJarError(error, jarPath) {
   }
 }
 
-exports.getJavaCoords = getJavaCoords;
-exports.getJarPath = getJarPath;
-exports.getLocalName = getLocalName;
-exports.getSearchUrl = getSearchUrl;
-exports.isServiceInstalled = isServiceInstalled;
+exports.getServicePath = getServicePath;
 exports.installServiceIfRequired = installServiceIfRequired;
-exports.fileExists = fileExists;
-exports.setupDownload = setupDownload;
-exports.installJar = installJar;
-exports.handleJarError = handleJarError;
